requirements.txt
numpy==1.24.4
scipy==1.10.1
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
torch==1.11.0
torchvision==0.10.0+cu111
torchaudio==0.9.0
tqdm==4.65.0
opencv-python==4.7.0.72
pillow==9.4.0
imageio==2.31.1
h5py==3.7.0
pyyaml==6.0
huggingface-hub==0.19.0rc0
transformers==4.33.3
timm==0.9.2
wandb==0.15.4
grad-cam==1.4.6
albumentations==1.3.0
